perf(ci): parallelize staff XCUITest across 2 simulator clones (plan S3)
The 18-test XCUITest suite ran serially at ~12 m and is the deploy-gate
critical path now that backend shards sit at 4-7 m. XCTest distributes
work to parallel workers per-class, and PatientCommunicationRoutingUITests
held 11 of the 18 tests - two workers would have capped at -4 m without
a split.

- project.yml: HummingbirdUITests parallelizable: true
- ci.yml: -parallel-testing-enabled YES -parallel-testing-worker-count 2
  on the XCUITest invocation (unit suite stays serial); Q3 PR-lane retry
  flags unchanged
- Split the routing class 5/6 through a shared
  PatientCommunicationRoutingUITestCase base (no test methods, so XCTest
  skips it): routing mutations + exact-replay recovery stay in
  PatientCommunicationRoutingUITests; the 401/403/404/200-omission
  revocation purges move to PatientCommunicationPurgeUITests. All 11
  test-method names and bodies are unchanged; class sizes are now
  6/5/3/2/2 so two workers balance ~9/9.
