+++ hummingbird/iosApp/HummingbirdUITests/PatientCommunicationRoutingUITests.swift
@@ -1,25 +1,9 @@
 import XCTest
 
-final class PatientCommunicationRoutingUITests: XCTestCase {
-    private var app: XCUIApplication!
-
-    override func setUp() {
-        super.setUp()
-        continueAfterFailure = false
-        app = XCUIApplication()
-        app.launchArguments = ["-HBStaffCommunicationsUITest"]
-        app.launchEnvironment = [
-            "HB_STAFF_COMM_UI_TEST": "1",
-            "HB_STAFF_COMM_UI_SCENARIO": "routing",
-        ]
-    }
-
-    override func tearDown() {
-        app.terminate()
-        app = nil
-        super.tearDown()
-    }
-
+/// Routing mutation journeys: reassign/reroute flows, their explicit
+/// confirmations, and exact-replay recovery for lost-committed requests.
+/// Revocation/purge journeys live in `PatientCommunicationPurgeUITests`.
+final class PatientCommunicationRoutingUITests: PatientCommunicationRoutingUITestCase {
     func testReassignRequiresBoundedSelectionsAndExplicitConfirmation() {
         app.launch()
         openRoutingAction("patientCommunications.routing.reassignButton")
@@ -122,111 +106,6 @@ final class PatientCommunicationRoutingUITests: XCTestCase {
         capture("patient-communications-routing-first-success-purged")
     }
 
-    func testInboxPolling401PurgesOpenDraftAndRequiresSignIn() {
-        let secret = "SESSION LOSS DRAFT 7391"
-        launchScenario("inbox_401_detail", seededDraft: secret)
-        openDetail()
-        assertSeededDraft(secret)
-
-        XCTAssertTrue(app.buttons["Sign in"].waitForExistence(timeout: 20))
-        XCTAssertFalse(app.textViews["patientCommunications.replyEditor"].exists)
-        XCTAssertFalse(elementContaining(secret).exists)
-        XCTAssertFalse(app.buttons["Confirm reroute"].exists)
-        capture("patient-communications-inbox-401-purged")
-    }
-
-    func testInboxPolling403And404PurgeOpenDraftThreadAndCommands() {
-        for (scenario, secret) in [
-            ("inbox_403_detail", "INBOX FORBIDDEN DRAFT 2648"),
-            ("inbox_404_detail", "INBOX NOT FOUND DRAFT 7315"),
-        ] {
-            launchScenario(scenario, seededDraft: secret)
-            openDetail()
-            XCTAssertTrue(
-                app.buttons["patientCommunications.routing.rerouteButton"].waitForExistence(timeout: 5),
-                scenario
-            )
-            assertSeededDraft(secret)
-
-            let unavailable = app.descendants(matching: .any)["patientCommunications.threadUnavailable"]
-            XCTAssertTrue(unavailable.waitForExistence(timeout: 20), scenario)
-            XCTAssertFalse(app.textViews["patientCommunications.replyEditor"].exists, scenario)
-            XCTAssertFalse(elementContaining(secret).exists, scenario)
-            XCTAssertFalse(app.buttons["patientCommunications.routing.rerouteButton"].exists, scenario)
-            XCTAssertFalse(app.buttons["patientCommunications.routing.retryExactButton"].exists, scenario)
-            XCTAssertFalse(app.buttons["Sign in"].exists, scenario)
-            capture("patient-communications-\(scenario)-purged")
-            app.terminate()
-        }
-    }
-
-    func testInboxPolling200OmissionPurgesOpenDraftThreadAndCommands() {
-        let secret = "INBOX OMITTED DRAFT 6184"
-        launchScenario("inbox_200_empty_detail", seededDraft: secret)
-        openDetail()
-        XCTAssertTrue(app.buttons["patientCommunications.routing.rerouteButton"].waitForExistence(timeout: 5))
-        assertSeededDraft(secret)
-
-        let unavailable = app.descendants(matching: .any)["patientCommunications.threadUnavailable"]
-        XCTAssertTrue(unavailable.waitForExistence(timeout: 20))
-        XCTAssertFalse(app.textViews["patientCommunications.replyEditor"].exists)
-        XCTAssertFalse(elementContaining(secret).exists)
-        XCTAssertFalse(app.buttons["patientCommunications.routing.rerouteButton"].exists)
-        XCTAssertFalse(app.buttons["patientCommunications.routing.retryExactButton"].exists)
-        XCTAssertFalse(app.buttons["Sign in"].exists)
-        capture("patient-communications-inbox-200-omission-purged")
-    }
-
-    func testCandidate401RefreshPurgesDraftAndRequiresSignIn() {
-        let secret = "CANDIDATE AUTH DRAFT 4826"
-        launchScenario("candidate_401_refresh", seededDraft: secret)
-        openDetail()
-        XCTAssertTrue(app.buttons["patientCommunications.routing.rerouteButton"].waitForExistence(timeout: 5))
-        assertSeededDraft(secret)
-        triggerDetailRefresh()
-
-        XCTAssertTrue(app.buttons["Sign in"].waitForExistence(timeout: 12))
-        XCTAssertFalse(app.textViews["patientCommunications.replyEditor"].exists)
-        XCTAssertFalse(elementContaining(secret).exists)
-        XCTAssertFalse(app.buttons["patientCommunications.routing.rerouteButton"].exists)
-    }
-
-    func testCandidate404RefreshPurgesDraftThreadAndCommands() {
-        let secret = "CANDIDATE DENIAL DRAFT 1564"
-        launchScenario("candidate_404_refresh", seededDraft: secret)
-        openDetail()
-        XCTAssertTrue(app.buttons["patientCommunications.routing.rerouteButton"].waitForExistence(timeout: 5))
-        assertSeededDraft(secret)
-        triggerDetailRefresh()
-
-        let unavailable = app.descendants(matching: .any)["patientCommunications.threadUnavailable"]
-        XCTAssertTrue(unavailable.waitForExistence(timeout: 12))
-        XCTAssertFalse(app.textViews["patientCommunications.replyEditor"].exists)
-        XCTAssertFalse(elementContaining(secret).exists)
-        XCTAssertFalse(app.buttons["patientCommunications.routing.rerouteButton"].exists)
-        XCTAssertFalse(app.buttons["patientCommunications.routing.retryExactButton"].exists)
-        capture("patient-communications-candidate-404-purged")
-    }
-
-    func testDetail403And404RefreshPurgeDraftBeforeUnavailableState() {
-        for (scenario, secret) in [
-            ("thread_403_refresh", "DETAIL FORBIDDEN DRAFT 8327"),
-            ("thread_404_refresh", "DETAIL NOT FOUND DRAFT 9053"),
-        ] {
-            launchScenario(scenario, seededDraft: secret)
-            openDetail()
-            assertSeededDraft(secret)
-            triggerDetailRefresh()
-
-            let unavailable = app.descendants(matching: .any)["patientCommunications.threadUnavailable"]
-            XCTAssertTrue(unavailable.waitForExistence(timeout: 12), scenario)
-            XCTAssertFalse(app.textViews["patientCommunications.replyEditor"].exists, scenario)
-            XCTAssertFalse(elementContaining(secret).exists, scenario)
-            XCTAssertFalse(app.buttons["patientCommunications.routing.retryExactButton"].exists, scenario)
-            app.terminate()
-        }
-    }
-
     func testLostCommittedRerouteOffersOnlyExplicitExactReplay() {
         app.launchEnvironment["HB_STAFF_COMM_UI_SCENARIO"] = "ambiguous_reroute"
         app.launch()
@@ -314,93 +193,5 @@ final class PatientCommunicationRoutingUITests: XCTestCase {
             "Exact replay must not append a second patient-visible reply"
         )
         capture("patient-communications-reply-exact-replay-confirmed")
-    }
-
-    private func openRoutingAction(_ identifier: String) {
-        openDetail()
-
-        let action = app.buttons[identifier]
-        for _ in 0..<8 where !action.exists {
-            app.swipeUp()
-        }
-        XCTAssertTrue(action.waitForExistence(timeout: 5))
-        reveal(action)
-        XCTAssertTrue(action.isHittable)
-        action.tap()
-    }
-
-    private func openDetail() {
-        let messages = app.tabBars.buttons["Messages"]
-        XCTAssertTrue(messages.waitForExistence(timeout: 5))
-        messages.tap()
-
-        let row = app.descendants(matching: .any)[
-            "patientCommunications.row.11111111-1111-4111-8111-111111111111"
-        ]
-        XCTAssertTrue(row.waitForExistence(timeout: 5))
-        row.tap()
-    }
-
-    private func launchScenario(_ scenario: String, seededDraft: String? = nil) {
-        if app.state != .notRunning { app.terminate() }
-        app = XCUIApplication()
-        app.launchArguments = ["-HBStaffCommunicationsUITest"]
-        app.launchEnvironment = [
-            "HB_STAFF_COMM_UI_TEST": "1",
-            "HB_STAFF_COMM_UI_SCENARIO": scenario,
-        ]
-        if let seededDraft {
-            app.launchEnvironment["HB_STAFF_COMM_UI_SEEDED_DRAFT"] = seededDraft
-        }
-        app.launch()
-    }
-
-    private func enterDraft(_ text: String) {
-        let editor = app.textViews["patientCommunications.replyEditor"]
-        XCTAssertTrue(editor.waitForExistence(timeout: 5))
-        editor.tap()
-        editor.typeText(text)
-        let typedValue = XCTNSPredicateExpectation(
-            predicate: NSPredicate(format: "value CONTAINS %@", text),
-            object: editor
-        )
-        XCTAssertEqual(
-            XCTWaiter.wait(for: [typedValue], timeout: 2),
-            .completed,
-            "Expected the reply editor to contain the exact test body; actual value: \(String(describing: editor.value))"
-        )
-    }
-
-    private func assertSeededDraft(_ text: String) {
-        let editor = app.textViews["patientCommunications.replyEditor"]
-        XCTAssertTrue(editor.waitForExistence(timeout: 5))
-        XCTAssertTrue((editor.value as? String)?.contains(text) == true)
-    }
-
-    private func triggerDetailRefresh() {
-        // These scenarios schedule a deterministic seven-second authorization refresh
-        // from the detail view. Do not tap an element that the refresh is allowed to
-        // remove between accessibility lookup and event synthesis; the state-specific
-        // assertion immediately after this helper is the actual completion barrier.
-        _ = app.navigationBars["Patient conversation"].waitForExistence(timeout: 1)
-    }
-
-    private func elementContaining(_ text: String) -> XCUIElement {
-        app.descendants(matching: .any)
-            .matching(NSPredicate(format: "label CONTAINS[c] %@", text))
-            .firstMatch
-    }
-
-    private func reveal(_ element: XCUIElement) {
-        for _ in 0..<8 where !element.isHittable {
-            app.swipeUp()
-        }
-    }
-
-    private func capture(_ name: String) {
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = name
-        attachment.lifetime = .keepAlways
-        add(attachment)
     }
 }
